hw_1/pyproject.toml
[project]
name = "break-into-data-pv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.72.6",
    "python-dotenv>=1.1.0",
]
